Container moved to SearchResults styled
